refactor: Move JSON decoding to API
Moves the `Jason.decode` call into the `API` module using HTTPoison's `process_response_body`.

This simplifies the application a bit by not having to decode, and also simplifies the stubbed API responses in `config_cat_test.exs` by not having to encode.

There's a couple of tests that get back an HTML 404 response (from nginx), even after specifying an `Accept` header, so I added protection against that and return the body unmodified in that case.

I also realized that there was no need to specify a `Content-Type` header because we only make `GET` requests and those don't have content.