fix: eliminate race conditions in tests
When @igorescobar tried to run Erlang 22 in CI, a couple of tests failed consistently.

These tests were both starting ConfigCat with an auto refresh policy. The auto refresh ended up happening before we could define the `stub` response to the `GET` request it was making, resulting in the failure.

One of the tests didn't need an auto fetch policy, so I switched it to manual.  For the other, I moved the `stub` definition prior to the `start_config_cat` call as it is in the other auto policy tests.