feat: create a ConfigCache behaviour and InMemory implementation
As with the Ruby version, we'll generate a cache key based on the config filename and SDK key when we hook this in.

In order to be able to have one cache per ConfigCat instance, we'll use the cache key as the GenServer name (after converting to an atom).  Each InMemoryCache instance will store the config as its state and we'll use the key to communicate with the correct GenServer.