refactor: use pluggable cache in client
We now allow the calling application to specify a `cache_api` to replace our default `InMemoryCache`.

The calling application is responsible for adding its cache to its own supervision tree.

If no cache_api is provided, we supply our own default `InMemoryCache`, adding it to our supervision tree.

To ensure that each ConfigCat instance has its own cache, we use the `cache_key` (after converting to an atom) as the GenServer name.