refactor: use atoms for data governance
As I was working on the Hex docs, I found that the DataGovernance module was painful to document.  I realized that we could replace it by using atom values instead, as we don't read numeric values from the configuration file anywhere.

This eliminates the `DataGovernance` module and uses the `:global` and `:eu_only` atoms instead where necessary.