refactor: extract CachePolicy behaviour to private module
When working on the Hex docs, I found that ex_doc was documenting CachePolicy as a behaviour and I couldn't hide that part of the docs, even though we want it to be private.

This extracts the callbacks and private types to a separate `CachePolicy.Behaviour` module to avoid the problem.

This is actually a bit clearer now, because CachePolicy was serving as both the behaviour for the implementations and as a public-facing facade for working with cache policies.  Now, each responsibility is in its own module.