🔧 Enable the MultiAlias rule
While the shorthand alias syntax takes up fewer lines of code, it's much harder to find all references to a module by searching for the fully-qualified module name. It also makes it more tedious to add or remove aliases especially when adding a second alias to a group, or removing the second-last alias in a group.