➕ Add Credo linter (#85)
* ♻️ Alphabetize dependencies

* ➕ Add credo for linting

- Generate default config
- Fix issues reported on first run

* 🔧 Enable credo strict mode

Fix the only new issue reported.

* 🔧 Enable some additional credo rules

* 🔧 Enable PassAsyncInTestCases rule

Fix one violation.

* 🔧 Enable StrictModuleLayout rule

Add .credo.exs to formatter config to ensure that it formats properly.

* 🏷️ Enable Specs rule and add missing specs

* 🔧 Enable the MultiAlias rule

While the shorthand alias syntax takes up fewer lines of code, it's much harder to find all references to a module by searching for the fully-qualified module name. It also makes it more tedious to add or remove aliases especially when adding a second alias to a group, or removing the second-last alias in a group.