♻️ Inline Client's public API into ConfigCat
Every time we added new public API to ConfigCat, we had to add it to both ConfigCat and ConfigCat.Client.

This inlines the `GenServer.call/cast` calls up into ConfigCat to eliminate this extra layer of indirection.

It's not the normal way of doing things in Elixir, but in this case it seems to make sense to do it this way.