♻️ Return only settings from OverrideDataSource
Simplify OverrideDataSource implementations to return only the flags/settings from the config, and to return an empty map if the settings can't be found/loaded.

This fixes a bug: if an override file couldn't be found, we would act as if we had no config at all, but we should still be falling back to the fetched config instead.