🔧 Catch compiler warnings
Set default options for the Elixir compiler in various environments to catch compiler warnings.

In dev and test, we allow compiler warnings (to allow running and testing incomplete code), but we make sure to continuously report those errors even when the affected files are not recompiled. That way, warnings don't appear once, get forgotten about, then disappear.

In production, we treat all warnings as errors.

In CI, we compile in the dev environment with `--warnings-as-errors` and also run the tests with the same flag.

Also, fix one compiler warning that snuck through from an earlier commit.