✅ Add more auto/lazy tests (#95)
Add tests to ensure that initial fetches don't happen if the cache is already populated with a recent entry.

The new tests helped me realize that I had the logic backward in auto mode, so I fixed that.

The fix made the offline test start to fail, so I simplified/refactored it a bit. That helped me realize that we weren't setting the online flag in the state before (possibly) re-fetching the config while going online, so the fetch wasn't actually happening.

These tests are similar in spirit to those in the other SDKs, but are unit tests that don't need as many sleeps/delays.