♻️ Make fetch logic consistent with other platforms
- Refresh the fetch time of the cache entry when a fetch returns 304
- Refresh the fetch time of the cache entry when a fetch returns a non-transient error (to reduce the load on the server)
- Make log messages consistent with other platforms

As I was doing this work, I realized that we were leaking HTTPoison.Error and HTTPPoison.Response structs out to callers of `force_refresh`. Instead, we now return an `{:error, message}` tuple.

I also noticed that, in the Ruby version, `force_refresh` returns an error message when offline which the Elixir version wasn't, so I fixed that.

I also noticed that the Ruby version does not return an error from force_refresh if there is a previously cached result. That doesn't seem right to me, so I did not make that change here.