✨ Add max_init_wait_time_seconds setting
In order to make this work correctly, I had to significantly modify how the Auto policy worked. Now, it always fetches in the background.

When a caller calls `get` before the initial fetch is completed, the caller is added to a list and we use `{:noreply, ...}` to block the caller while still allowing the GenServer to keep working.

Once the initial fetch completes or times out, all pending callers are unblocked with `GenServer.reply`.

See [my blog post](https://blog.sequin.io/genserver-reply-dont-call-us-well-call-you/) for more details on this technique.