⚡️ Respond to all queued fetch requests with single response
When performing a fetch, queue up any new requests and then respond to all of the fetches with the result of the pending fetch.

This avoids making multiple consecutive fetch requests to the server.

Uses a technique I describe in [this blog post](https://blog.sequin.io/genserver-reply-dont-call-us-well-call-you/](https://blog.sequin.io/genserver-reply-dont-call-us-well-call-you/) along with a list of pending callers in the GenServer state.