♻️ Consistent log messages
Make log messages consistent with other language SDKs, including the addition of `event_id`'s.

Discovered that `:local_only` mode would cause errors when calling certain functions, so added a new `CachePolicy.Null` for that case.

Updated the typespec for `ConfigCache.set` to allow for it to return an error tuple so that we can log the error.

Note that the accumulation of evaluation log messages will be done as part of another ticket.