✨ Report a single log of evaluation steps
Start an Agent to accumulate evaluation log messages in a list, then print a single message at the end with those messages all newline separated.

Note that, contrary to the Ruby SDK, I've continued to log these at the `debug` level because in the past we've had complaints about how noisy these logs get.

If that is not OK, I can change it.