✨ Ensure that only one instance can use the same SDK key (#108)
* ✨ Ensure that only one instance can use the same SDK key

- Switch the Supervisor to use a via_tuple with an extra "value" argument of the SDK key.
- When starting a new Supervisor, first check that there are no other supervisors already using that SDK key. If there are, raise an error.

The other SDKs return the matching instance after logging, but that's not really possible to do in an Elixir supervision tree, so instead we fail hard.

I've made some changes to the log message to match that reality, so it doesn't quite conform to the standardized log messages.

* ✅ Avoid conflict with integration tests

The using block test was failing frequently due to a conflict with the integration tests. Both were trying to start ConfigCat instances with the same SDK Key which is no longer allowed.