✅ Fix flaky tests
For a while now, I've been seeing spurious test failures, but hadn't been able to figure out the root cause.

It turns out that clearing the entire `InMemoryCache` before tests was causing test interference. This is obvious in retrospect.

There's also a spurious failure in CI with the integration tests where a second instance tries to start with the same SDK key as another. I'm not sure if I've fixed that one with these changes.

Changes:
- Clear only a specific cache key from the InMemoryCache, and only when multiple tests will use the same key.
- Mark the integration tests as `async: false` to be sure they run in isolation.
- Eliminate log noise from the tests.