✅ Add attribute value conversion tests
In order to test non-UTC timezones, we need a timezone database. I've added one for the `test` environment only and I make it the default in `test_helper.exs`.

In order to pass the tests, I had to adjust the implementation of `to_unix_seconds` to return fractional seconds.

Note that Elixir does not support +inf/-inf/NaN float values, so I had to leave those tests commented out.