♻️ Minor cleanups to logging code
- Reverse names of `log_evaluating_condition_result` and `log_evaluating_condition_final_result` to better reflect their usage

- Move condition_count check to caller for `log_evaluating_condition_result`

- Move `log_evaluating_targeting_rules` to its proper alphabetical position

♻️ Refactor log_evaluating_condition_final_result

Pass a `newline?` param instead of `condition_count` to prepare for a better calculation of `newline?`.  Involved moving some increase_/decrease_indent calls around.

✨ Pass newline_before_then? flag

Based on the final evaluated condition, determine whether or not there should be a newline before the `THEN` in the logs.

This was a bit tricky to properly accumulate and pass around.